Bug in no-case fixed
The comparisons for updating the upper and lower limits where inversed. These have been fixed.
Furthermore, a check for whether both limits are present have been added. If both limits are present that means the estimate splits into two ranges which is not currently supported and thus will not be used to update the estimate.